Move auth to utils
